Реализует прием данных и генерацию разметки на их основе для точки маршрута
