+++ src/view/creating-form.js
@@ -1,4 +1,5 @@
 export const createCreatingFormTemplate = () => {
+
   return `<li class="trip-events__item">
     <form class="event event--edit" action="#" method="post">
       <header class="event__header">
